+++ ligo/skymap/bayestar/filter.py
@@ -26,12 +26,6 @@ import numpy as np
 from scipy import fftpack as fft
 from scipy import interpolate, linalg
 
-try:
-    from numpy import trapezoid
-except ImportError:
-    # FIXME: Remove after we require Numpy >=2.0.0.
-    from numpy import trapz as trapezoid
-
 log = logging.getLogger("BAYESTAR")
 
 
@@ -573,7 +567,7 @@ class SignalModel:
         h = h.data.data[first_nonzero : last_nonzero + 1]
 
         self.denom_integrand = 4 / (2 * np.pi) * h
-        self.den = trapezoid(self.denom_integrand, dx=self.dw)
+        self.den = np.trapezoid(self.denom_integrand, dx=self.dw)
 
     def get_horizon_distance(self, snr_thresh=1):
         return np.sqrt(self.den) / snr_thresh
@@ -582,7 +576,7 @@ class SignalModel:
         """Get the average of a function of angular frequency, weighted by the
         signal to noise per unit angular frequency.
         """
-        num = trapezoid(func(self.w) * self.denom_integrand, dx=self.dw)
+        num = np.trapezoid(func(self.w) * self.denom_integrand, dx=self.dw)
         return num / self.den
 
     def get_sn_moment(self, power):