+++ ligo/skymap/extern/numpy/quantile.py
@@ -34,12 +34,7 @@
 import warnings
 
 import numpy as np
-
-try:
-    import numpy._core.numeric as _nx
-except ImportError:
-    # FIXME: Remove after we require Numpy >=2.0.0.
-    import numpy.core.numeric as _nx
+import numpy._core.numeric as _nx
 
 
 def _ureduce(a, func, **kwargs):